Add helper to verify API token
